settings page added, language selection removed from app bar
