+++ app/src/main/java/com/example/myfirstapp/ui/settings/SettingsActivity.java
@@ -19,10 +19,7 @@ import com.example.myfirstapp.ui.BaseActivity;
 import java.util.Locale;
 
 public class SettingsActivity extends BaseActivity {
-    private ListView listView;
-    private int digit[] = new int[] {1,2,3,4,5,6,7,8,9,10};
-    private ListAdapter listAdapter;
-    private Locale locale;
+
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,27 +29,10 @@ public class SettingsActivity extends BaseActivity {
         setTitle(getString(R.string.app_name));
         bottomNavigationView.setSelectedItemId(R.id.settings);
 
-        //Get phone default language
-        String lang = Locale.getDefault().getLanguage();
-        //not en or fr, set to en
-        if(!lang.equals("en") || !lang.equals("fr") ) lang = "en";
-        setAppLocale(lang);
+        getFragmentManager().beginTransaction()
+                .replace(R.id.flFragment,new SettingsFragment())
+                .commit();
 
-        //list view
-        listView = findViewById(R.id.listView);
-        LayoutInflater inflater = (LayoutInflater) this.getSystemService(this.LAYOUT_INFLATER_SERVICE);
-
-        //Add Header View
-        View headerView = inflater.inflate(R.layout.listview_header, null, false);
-        TextView headerText = headerView.findViewById(R.id.listHeader);
-        headerText.setText("Settings");
-        listView.addHeaderView(headerView);//Add view to list view as header view
-
-        listAdapter = new ListAdapter(this,digit);
-        listView.setAdapter(listAdapter);
-
-        // Set default selection
-        //bottomNavigationView.setSelectedItemId(R.id.nav_none);
 
     }
     @Override
@@ -78,17 +58,6 @@ public class SettingsActivity extends BaseActivity {
         alertDialog.show();
     }
 
-    private void setAppLocale(String language){
-        locale = new Locale(language);
-        Resources res = getResources();
-        Configuration config = res.getConfiguration();
-        //DisplayMetrics displayMetrics = res.getDisplayMetrics();
-
-        Locale.setDefault(locale);
-        config.setLocale(locale);
-        getBaseContext().getResources().updateConfiguration(config, getBaseContext().getResources().getDisplayMetrics());
-
-    }
     @Override
     public void onConfigurationChanged(Configuration newConfig) {
         super.onConfigurationChanged(newConfig);
@@ -99,18 +68,6 @@ public class SettingsActivity extends BaseActivity {
     @Override
     public boolean onOptionsItemSelected(MenuItem item){
         int id = item.getItemId();
-
-        if (id == R.id.btnLang) {
-            String lang = locale.getLanguage();
-            Log.d("language",lang);
-            String newLang = "en";
-            if(lang.equals("en")) newLang = "fr";
-            else if(lang.equals("fr")) newLang = "en";
-
-            Log.d("new language",newLang);
-            setAppLocale(newLang);
-            return true;
-        }
 
         return super.onOptionsItemSelected(item);
