+++ app/src/main/java/com/example/myfirstapp/ui/settings/SettingsFragment.java
@@ -1,0 +1,16 @@
+package com.example.myfirstapp.ui.settings;
+
+import com.example.myfirstapp.R;
+import android.os.Bundle;
+import android.preference.PreferenceFragment;
+
+public class SettingsFragment extends PreferenceFragment{
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        // Load the preferences from an XML resource
+        addPreferencesFromResource(R.xml.preferences);
+    }
+}
