+++ app/src/main/res/menu/menu_top.xml
@@ -2,12 +2,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto">
 
-
-    <item
-        android:id="@+id/btnLang"
-        android:title="@string/language"
-        android:icon="@drawable/ic_lang_foreground"
-        app:showAsAction="ifRoom|withText" />
     <item
         android:id="@+id/btnAbout"
         android:title="@string/about"
