--- app/src/main/res/mipmap-anydpi-v26/ic_lang.xml
@@ -1,5 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<adaptive-icon xmlns:android="http://schemas.android.com/apk/res/android">
-    <background android:drawable="@drawable/ic_lang_background"/>
-    <foreground android:drawable="@drawable/ic_lang_foreground"/>
-</adaptive-icon>
