+++ app/src/main/res/values/array.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string-array name="pref_lang_entries">
+        <item>English</item>
+        <item>Francais</item>
+    </string-array>
+    <string-array name="pref_lang_values">
+        <item>en</item>
+        <item>fr</item>
+    </string-array>
+</resources>
