+++ app/src/main/res/xml/pref_headers.xml
@@ -1,0 +1,10 @@
+<preference-headers xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <!-- These settings headers are only used on tablets. -->
+
+    <header
+        android:fragment="com.example.myfirstapp.ui.settings.SettingsActivity$AboutPreferenceFragment"
+        android:icon="@drawable/ic_about_foreground"
+        android:title="@string/preferenceAboutHeader" />
+
+</preference-headers>
